+++ mayan/apps/views/icons.py
@@ -24,3 +24,8 @@ icon_remove_all = Icon(
 )
 icon_sort_down = Icon(driver_name='fontawesome', symbol='sort-down')
 icon_sort_up = Icon(driver_name='fontawesome', symbol='sort-up')
+
+
+icon_user_view_modes = Icon(
+    driver_name='fontawesome', symbol='th-large'
+)