+++ HISTORY.rst
@@ -96,6 +96,7 @@
   or full page load.
 - Added persistent view mode setting. This features stores the user selected
   view mode for a specific view.
+- Add view to show the detail of each user view modes.
 
 4.7.1 (2024-06-04)
 ==================